+++ CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package ariles_ros
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+* Use Eigen Cholesky decomposition instead of a custom one, throw on
+  semi-definite Hessians.
+
+
 1.0.2 (2019-12-31)
 ------------------